requirements.txt
discord.py==2.3.2
langchain==0.1.19
langchain_core==0.1.52
langchain_openai==0.1.6
Requests==2.31.0
